Moving even more into the function
